% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Academic Theme in addition to ggplot2}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{A positive integer value}
}
\value{
plot with theme added when passed with ggplot2 (centralized plot title, grid lines removed)
}
\description{
Academic Theme in addition to ggplot2
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars) + geom_point(aes(x = mpg, y = hp)) + labs(title = "plot_title")
p + theme_academic()
}
